Разделение файла main.dart на несколько
